pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "parakh_ai"
version = "1.0.0"
description = "ParakhAI: Few-Shot Visual Anomaly Detection for Indian Micro-Manufacturing SMEs"
authors = [{name = "ParakhAI Team"}]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "anomalib>=1.1.0",
    "faiss-cpu>=1.7.4",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "albumentations>=1.3.0",
    "imagehash>=4.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "sse-starlette>=1.6.5",
    "aiosqlite>=0.19.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools]
packages = ["parakh_ai"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
